refactor: apply layer-discipline review fixes
Layer breach fixes:
- contributor.py: replace click.prompt with chooser callback, sys.exit with
  ContributeError; CLI defines the click.prompt-based chooser inline
- orchestrator.py: raise BeaconSyncError instead of silently skipping on
  NEEDS_CONFIRMATION (non-interactive conflict); restore exit-1 behavior
- domains/artifact/agent.py: raise/exit on NEEDS_CONFIRMATION instead of
  silently skipping; aligns handle_install_agent with the same pattern

Structural cleanup:
- Move BeaconSyncError from orchestrator.py to core/exceptions.py
- Add ContributeError to core/exceptions.py
- Remove dead SyncOrchestrationResult fields: has_synced_contexts,
  gitignore_updated, agents_synced (set but never read by cli/sync.py)
- Remove dead imports build_agents_paths, build_skills_paths from orchestrator.py
- Remove duplicate deferred detect_agents import inside skill.py:build_skills_paths
- cli/sync.update: restore reset_cmd.callback(project=project) delegation
  (removes copy-paste); drop dead oc_added/cc_added assignments in sync handler

Migration fixes (pre-existing branch work):
- adoption/adopter.py split into apply.py, discovery.py, models.py, tui.py
- check_warehouse_git_clean/on_main_branch moved to core/git_health.py
  returning GitHealthResult; callers updated
- handle_soft_block removed from utils/display.py; replaced by resolve_conflict
  in utils/interaction.py returning ConflictResolution (StrEnum)
- reset.py: replace sys.exit with ResetError; remove_artifacts_dir returns
  Path | None instead of bool; hoist deferred imports to module level

Test fixes:
- Update test mocks from old string-based git-check API to GitHealthResult
- Fix is_interactive patch path: orchestrator -> cli.sync
- Fix adopt test imports following adopter.py split

Architecture:
- TC10: forbid click/rich/sys.exit in domains/ and core/ with explicit
  per-file waivers documenting known violations pending cleanup
- spec: update cross-domain import rule to depth <= 4 (empty __init__.py
  forces submodule imports); add TC10 scenario to spec
