+++ openspec/changes/introduce-domain-layer/specs/layered-architecture/spec.md
@@ -45,10 +45,19 @@ Imports between layers SHALL follow one direction only: `cli → domains → cor
 - **WHEN** `pytest libs/beacon/tests/unit/test_architecture.py` runs
 - **THEN** it SHALL fail if any module under `beacon/utils/` contains a `from beacon.cli`, `from beacon.domains`, or `from beacon.core` import
 
-#### Scenario: Cross-domain imports go through top-level modules
+#### Scenario: Cross-domain imports are bounded by depth
 
 - **WHEN** a module in `beacon/domains/<A>/` imports from `beacon/domains/<B>/`
-- **THEN** the import path MUST reference a module directly under `beacon/domains/<B>/`, not a deeper internal module
+- **THEN** the import path MUST be at most four parts deep: `beacon.domains.<B>.<module>`
+- **NOTE** Because `__init__.py` files are empty (no re-exports), imports target submodules
+  directly rather than the package. The constraint is therefore depth ≤ 4, not "package only".
+  Imports five or more parts deep (e.g. `beacon.domains.B.sub.nested`) are forbidden.
+
+#### Scenario: TC5 enforces the depth limit
+
+- **WHEN** `pytest libs/beacon/tests/unit/test_architecture.py` runs
+- **THEN** it SHALL fail if any cross-domain import (`from beacon.domains.<B>.*`) exceeds
+  four path components
 
 ### Requirement: Bounded contexts
 
@@ -141,3 +150,19 @@ The repository SHALL include `libs/beacon/tests/unit/test_architecture.py` which
 
 - **WHEN** a contributor runs `pytest` from the repository root
 - **THEN** `test_architecture.py` SHALL be collected and executed without requiring extra flags or markers
+
+### Requirement: Domain and core layers are CLI-framework-free
+
+Modules under `beacon/domains/` and `beacon/core/` SHALL NOT import `click`, `rich`, or call
+`sys.exit(...)`. Interactive UX (prompts, confirmations, console output, process termination)
+belongs exclusively in `beacon/cli/`.
+
+Domain functions that need caller input SHALL accept callback parameters or return structured
+result objects; CLI handlers own the UX.
+
+#### Scenario: TC10 enforces the no-CLI-in-domains rule
+
+- **WHEN** `pytest libs/beacon/tests/unit/test_architecture.py` runs
+- **THEN** it SHALL fail if any module under `beacon/domains/` or `beacon/core/` imports
+  `click` or `rich`, or calls `sys.exit(...)`, unless the file has an explicit entry in
+  the `_TC10_WAIVERS` dict (which documents known violations pending cleanup)
